+++ src/AOloopControl/AOloopControl_ProcessModeCoefficients.h
@@ -11,18 +11,18 @@ typedef struct
 
     char DMmodesname[80];
      // BLOCKS OF MODES
-    uint_fast16_t DMmodesNBblock;             /**< number of mode blocks (read from parameter) */
-    uint_fast16_t NBmodes_block[100];         /**< number of modes within each block (computed from files by AOloopControl_loadconfigure) */
-    uint_fast16_t modeBlockIndex[MAXNBMODES]; /**< block index to which each mode belongs (computed by AOloopControl_loadconfigure) */
-    uint_fast16_t indexmaxMB[maxNBMB]; 
-	uint_fast16_t NBDMmodes;
+    uint32_t DMmodesNBblock;             /**< number of mode blocks (read from parameter) */
+    uint32_t NBmodes_block[100];         /**< number of modes within each block (computed from files by AOloopControl_loadconfigure) */
+    uint32_t modeBlockIndex[MAXNBMODES]; /**< block index to which each mode belongs (computed by AOloopControl_loadconfigure) */
+    uint32_t indexmaxMB[maxNBMB]; 
+	uint32_t NBDMmodes;
 
 
 
 	 // COMPUTED BY OPEN LOOP RETRIEVAL PROCESS
     double           RMSmodes;
     double           RMSmodesCumul;
-    uint_fast64_t    RMSmodesCumulcnt;
+    uint64_t    RMSmodesCumulcnt;
 
 
 	// block statistics (instantaneous)
@@ -41,7 +41,7 @@ typedef struct
 	
 	
 	// averaged
-	uint_fast32_t AveStats_NBpt;       // averaging interval
+	uint32_t AveStats_NBpt;       // averaging interval
 	double blockave_PFresrms[100];     // open loop RMS
 	double blockave_OLrms[100];        // open loop RMS
 	double blockave_Crms[100];         // correction RMS
